pyproject.toml
[tool.poetry]
name = "code-ml4cps-paper"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [
    {include = "pipeline"},
    {include = "container_component_src"}
]


[tool.poetry.dependencies]
python = ">=3.9,<3.12.0"
jupyter = "^1.0.0"
ipython = "^8.18.1"
ipykernel = "^6.27.1"
kfp = "^2.4.0"
s3fs = "^2023.12.2"
toml = "^0.10.2"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
click = "^8.1.7"
pyarrow = "^14.0.1"
pandas = "^2.1.4"
loguru = "^0.7.2"
dask = "^2023.12.0"
dask-kubernetes = "^2023.10.0"
bokeh = "^3.3.2"
xlrd = "^2.0.1"
scikit-learn = "^1.3.2"
plotly = "^5.18.0"
torch = "^2.1.2"
pytorch-lightning = "^2.1.2"

[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"
pdbpp = "^0.10.3"
matplotlib = "^3.8.2"
nbstripout = "^0.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
venvPath = "."
venv = ".venv"
